+++ python_modules/dagster/dagster/core/config_types.py
@@ -1,6 +1,8 @@
 import re
 from dagster import check
 
+from dagster.utils import camelcase
+
 from .config import (
     Context,
     Environment,
@@ -347,15 +349,3 @@ class ExecutionConfigType(DagsterCompositeType):
 
     def evaluate_value(self, value):
         return process_incoming_composite_value(self, value, lambda val: Execution(**val))
-
-
-# Adapted from https://github.com/okunishinishi/python-stringcase/blob/master/stringcase.py
-def camelcase(string):
-    string = re.sub(r'^[\-_\.]', '', str(string))
-    if not string:
-        return string
-    return str(string[0]).upper() + re.sub(
-        r'[\-_\.\s]([a-z])',
-        lambda matched: str(matched.group(1)).upper(),
-        string[1:],
-    )
