+++ python_modules/dagster/dagster/utils/utils_tests/test_camelcase.py
@@ -1,0 +1,8 @@
+from dagster.utils import camelcase
+
+
+def test_camelcase():
+    assert camelcase('foo') == 'Foo'
+    assert camelcase('foo_bar') == 'FooBar'
+    assert camelcase('foo.bar') == 'FooBar'
+    assert camelcase('foo-bar') == 'FooBar'
